backend/requirements.txt
google-genai>=1.0
PyMuPDF>=1.24
python-dotenv>=1.0
requests>=2.28
supabase>=2.0
fastapi
uvicorn
python-multipart
pydantic
